Extract PriorityQueue class to a separate file
